Added some simple introspection functions.
These permit testing of explicit features, e.g.

    if (htslib_test_feature(HTS_FEATURE_PLUGINS)) {
        ...
    }

or capturing the entire feature bit-field via htslib_features().
(This function may be redundant so perhaps is a candidate for
culling?).

This also permits querying compilation details: CC, CFLAGS, LDFLAGS
and CPPFLAGS.

Finally there is a htslib_feature_string() function which is mainly
for verbose feature printing rather than programmatic parsing, but it
conveniently also permits us to get features from a library simply by
running "strings" on the binary.

Note this doesn't actually do any testing so it's not in the Makefile.
It's hard to know quite what testing it could do given by design the
output would differ based on how the user built it.  We could maybe
call it to make sure it doesn't crash, but that's not likely to be
something we need regression testing on.