Further update the NEWS file.
Changes include expanding some sections with example usage, a bit of
reordering to bring primary changes to the top, removal of the fuzzing
section (a bit of git blaming shows all are fixes to code that didn't
exist in the previous release), and some minor line wrapping.

Some of this will inevitably be duplicated in the Samtools NEWS file,
specifically the explanations of using CRAM 3.1 and the filtering
syntax, but I think that's appropriate.

Updated introspection function names to match those in PR#1222